% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{ct}
\alias{ct}
\title{Calculate time difference}
\usage{
ct(name, units = "sec")
}
\arguments{
\item{name}{name of timer}

\item{units}{units to return in. One of \code{c("sec","min","hr")}}
}
\value{
user, system, and elapsed time.
}
\description{
Calculate the time differnce for a timer created by \code{st()}
and \code{et()}.
}
\examples{
st("test")
Sys.sleep(2)
et("test")
ct("test")

}
\seealso{
\code{\link{st}}, \code{\link{et}}
}

