% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/password_from_file.R
\name{password_from_file}
\alias{password_from_file}
\title{Create a password from a file}
\usage{
password_from_file(file, pwd_length = 3, combn_char = NA)
}
\arguments{
\item{file}{plaintext file to draw words from}

\item{pwd_length}{number of words to select for password}

\item{combn_char}{character to combine password with}
}
\description{
select a random set of words from a text file and combine into a password.
Project Gutenberg books, academic papers, and scripts all make great sources for
memorable passwords.
}

