% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.R
\name{st}
\alias{et}
\alias{st}
\title{Timing Functions}
\usage{
st(name)

et(name)
}
\arguments{
\item{name}{name of timer}
}
\value{
None
}
\description{
Wrapper functions for \code{proc.time()}. Creates variables called
\code{st_name} and \code{et_name}, respectivly
}
\examples{
st("test")
# Long running code here
Sys.sleep(2)
et("test")

}
\seealso{
\code{\link{ct}}
}

