% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{trim}
\alias{trim}
\title{Trim extra space around string}
\usage{
trim(string)
}
\arguments{
\item{string}{the string to be trimmed}
}
\value{
the trimmed substring
}
\description{
Remove any extra space around a string
}
\examples{
a <- " foo "
trim(a)
b <- "\\n foo \\n "
trim(b)

}

